Make test-apheleia.el work.
Since the callback of apheleia--run-formatter is run in an asynchronous
process, all output to stdout was being handled by the inferior process.
Output must be retrieved in the main process with accept-process-output.

Additionally, any formatter that uses the file directive will not run if
buffer doesn't match the file on disk. with-temp-buffer doesn't have an
associated file so most formatters wouldn't run.