Log an error when a formatter's executable is not found (#269)
I was trying to set up a new formatter and it wasn't in my exec-path, so
it wasn't running anything. But there was no message in the apheleia
logs, even with `apheleia-log-debug-info` enabled. This adds a log entry
when the executable is not found.

Hopefully the formatting is okay - I think Apheleia re-indented it so
hopefully that's good.

Fixes #268

---------

Co-authored-by: Radon Rosborough <radon@intuitiveexplanations.com>