Pass --text arg to diff (#356)
For some reason, GNU diff on Windows (from the MinGW project) fails to
diff the file contents when passed over stdin unless the `--text`
argument is passed. This argument forces diff to treat the inputs as
text, and not binary.

To reproduce this error with Bash on Windows:

```
$ echo foo > foo
$ echo bar > bar
$ cat bar | diff.exe foo -
diff.exe: -: Invalid argument
$ cat bar | diff.exe --text foo -
1c1
< foo
---
> bar
```

It seems like this flag is harmless in this case, since Apheleia will
always be used for diffing text files, and never binary files.