Add styleR for ess-r-mode (#377)
Hi! My first time adding a formatter, please let me know if I missed
anything.

Some things I'm unclear about:

1. How to fail gracefully when [styleR](https://styler.r-lib.org/)
package is absent from the system
2. Whether the formatter should be named `ess-r` or something more
general like `rlang`

## Test case

FYI, there are loads of tests to pick from on
https://github.com/r-lib/styler/tree/main/tests/testthat.

I picked
[unindention/mixed-in.R](https://github.com/r-lib/styler/blob/main/tests/testthat/unindention/mixed-in.R),
but oddly, the output
[unindention/mixed-out.R](https://github.com/r-lib/styler/blob/main/tests/testthat/unindention/mixed-out.R)
isn't quite the same as we got!

I don't know if that's an upstream issue, or that upstream changed style
since that test was written, or that perhaps the file is meant to be
concatenated into something else.