apheleia-npx: return silently when formatter is missing (#383)
Behavior of `aphaleia-npx` currently differs from other formatters by
triggering an error when formatter is missing (causing a message to be
displayed in minibuffer and an error buffer to be created).

This PR adds a test to check if command is on user's PATH before
executing it in order to make `aphaleia-npx` silently ends when the
formatter command is not found at project and system level.